+++ Package.swift
@@ -6,7 +6,7 @@ let package = Package(
     products: [
         .library(name: "TCP", targets: ["TCP"]),
         .library(name: "HTTP", targets: ["HTTP"]),
-        .library(name: "WebSockets", targets: ["WebSockets"]),
+        .library(name: "WebSocket", targets: ["WebSocket"]),
     ],
     dependencies: [
         // Core extensions, type-aliases, and functions that facilitate common tasks.
@@ -19,13 +19,13 @@ let package = Package(
         .package(url: "https://github.com/vapor/crypto.git", .revision("rework")),
     ],
     targets: [
-        .target(name: "Performance", dependencies: ["HTTP", "TCP", "WebSockets"]),
+        .target(name: "Performance", dependencies: ["HTTP", "TCP", "WebSocket"]),
         .target(name: "TCP", dependencies: ["Debugging", "Core", "libc"]),
         .testTarget(name: "TCPTests", dependencies: ["TCP"]),
         .target(name: "CHTTP"),
         .target(name: "HTTP", dependencies: ["CHTTP", "TCP"]),
         .testTarget(name: "HTTPTests", dependencies: ["HTTP"]),
-        .target(name: "WebSockets", dependencies: ["HTTP", "Core", "Crypto"]),
-        .testTarget(name: "WebSocketsTests", dependencies: ["WebSockets"]),
+        .target(name: "WebSocket", dependencies: ["HTTP", "Core", "Crypto"]),
+        .testTarget(name: "WebSocketTests", dependencies: ["WebSocket"]),
     ]
 )
